+++ transformer_lens/components/unembed.py
@@ -18,10 +18,10 @@ class Unembed(nn.Module):
         self.cfg = HookedTransformerConfig.unwrap(cfg)
         # Note that there's a separate variable for d_vocab_out and d_vocab (the input vocab size). For language tasks these are always the same, but for algorithmic tasks we may want them to be different.
         self.W_U: Float[torch.Tensor, "d_model d_vocab_out"] = nn.Parameter(
-            torch.empty(self.cfg.d_model, self.cfg.d_vocab_out, dtype=cfg.dtype)
+            torch.empty(self.cfg.d_model, self.cfg.d_vocab_out, dtype=self.cfg.dtype)
         )
         self.b_U: Float[torch.Tensor, "d_vocab_out"] = nn.Parameter(
-            torch.zeros(self.cfg.d_vocab_out, dtype=cfg.dtype)
+            torch.zeros(self.cfg.d_vocab_out, dtype=self.cfg.dtype)
         )
 
     def forward(
